#35: Allow specifying to use HTTPS and raise TLS errors for Buckets
