+++ doc/changes/changes_2.2.0.md
@@ -11,6 +11,7 @@ See the [user guide](https://github.com/exasol/bucketfs-java/blob/main/doc/user_
 ## Features
 
 * [#17](https://github.com/exasol/bucketfs-java/issues/17): Added support for creating new buckets
+* [#35](https://github.com/exasol/bucketfs-java/issues/35): Added support using self-signed certificates and ignoring TLS errors
 
 ## Dependency Updates
 
@@ -22,7 +23,7 @@ See the [user guide](https://github.com/exasol/bucketfs-java/blob/main/doc/user_
 
 ### Test Dependency Updates
 
-* Updated `com.exasol:exasol-testcontainers:3.5.3` to `5.0.0`
+* Updated `com.exasol:exasol-testcontainers:3.5.3` to `5.1.0`
 * Updated `org.mockito:mockito-junit-jupiter:3.11.2` to `3.12.4`
 
 ### Plugin Dependency Updates
